Extract src/usageview.ts — the usage & forecast markup
First of the render-module splits (PLAN Phase 1). No unit tests by design:
these functions build strings, and snapshotting template literals mostly
re-asserts itself.

Moved: the forecast presentation the footer popup and the Usage settings tab
share (foreText, verdictChip, usageRow) and the whole analytics panel — tiles,
the 12-month spend heatmap, the token bars, model/token mixes, top projects and
the two forecast cards. 268 lines out of main.ts.

The boundary is "does it touch anything but data": everything moved returns a
string and reads only ./usage, ./rl, ./format and accentFor, so the module needs
no seam back into main.ts at all — no hooks, unlike the last two slices. What
stayed behind is what pokes the DOM or the backend: renderUsagePop/openUsagePop
(they poke $() and coordinate with closeFootMenus) and refreshTokens (an async
scan that repaints the settings dialog). refreshTokens still owns the
`tokenScanning` flag the panels read to choose between a skeleton and a "no data"
line, so that follows state.ts's convention — a live binding to read,
setTokenScanning to write.

Sliced by exact line range rather than by hand: the moved block diffs
byte-identical against HEAD:src/main.ts apart from the added setter and the
`export` keywords on the four functions main.ts still calls.

Verified in the browser: the panel renders 37KB of markup / 699 nodes with every
section present, usageRow produces its verdict chip and forecast line, and the
data path is intact end to end (a seeded day reaches the Total spend tile).
